타입 수정

+++ backend/jobplatform/src/main/java/jobplatform/fo/user/controller/M_MypageController.java
@@ -79,7 +79,7 @@ public class M_MypageController {
 	
 	//마이페이지 개인정보 수정
 	@GetMapping("/infrmtn")
-	public ResponseEntity<Map<String, Object>> userInfrmtn(@RequestParam("mbr_sq") int mbr_sq) {
+	public ResponseEntity<Map<String, Object>> userInfrmtn(@RequestParam("mbr_sq") Long mbr_sq) {
 		System.out.println("어휴ㅡㅜㅜㅜ======="+mbr_sq);
 		Map<String, Object> userInfrmtn = myPageService.findResumeM(mbr_sq);
 		System.out.println(userInfrmtn);
